api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.0",
    "Pillow>=10.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "sqlmodel>=0.0.34",
    "uvicorn>=0.41.0",
]
